Hello world 5
Correcció Int per Long

+++ Math_is_fun_5/Math_is_fun_5.ino
@@ -9,7 +9,7 @@
 
 //******************************* VARIABLES ***************************************
 //define the variable
-int drive_gb = 100; //define 100 GB
+long drive_gb = 100; //define 100 GB
 long drive_mb; // change int for long to get more numbers
 //******************************** SETUP ******************************************
 void setup()
